Rearrange Rust tests for C
