+++ tests/README.md
@@ -1,0 +1,3 @@
+# rust tests
+
+Rust tests, including tests for calling legacy C code.
