+++ tests/c.rs
@@ -1,0 +1,5 @@
+// Copyright 2024 Martin Pool
+
+//! Tests from Rust into bound C code.
+
+mod c_tests;
