+++ tests/c_tests/mod.rs
@@ -1,0 +1,2 @@
+pub mod compress_c;
+pub mod trace;
